chore/Reorganized file structure, moved banner-[name] components to components instead of components/ui, created wip footer-link component
