+++ src/components/banner/banner-gradient/BannerGradient.tsx
