+++ src/components/banner/banner-image/BannerImage.stories.tsx
