+++ src/components/banner/banner-image/BannerImage.tsx
