+++ src/stories/ui/banner/samahan-members-banner.stories.tsx
@@ -1,4 +1,4 @@
-import { BannerContainer } from '@/components/ui/banner/banner-container/BannerContainer';
+import { BannerContainer } from '@/components/banner/banner-container/BannerContainer';
 import { BannerOverlay } from '@/components/ui/banner/banner-overlay/BannerOverlay';
 import { SAMAHANMembersBanner } from '@/components/ui/banner/samahan-members-banner/SAMAHANMembersBanner';
 import type { Meta, StoryObj } from '@storybook/react';
